rfctr: move filter test files in appropriate directory
